centralized layout file
